improve HUD and time taken entries in condition list

+++ Source/StudyFrameworkPlugin/Public/SFCondition.h
@@ -57,5 +57,5 @@ protected:
 
 private:
 
-	float StartTime;
+	double StartTime;
 };
